Separate LDAP bind, lookup, and authentication
The previous LDAP setup could not handle users in nested OUs. This
commit separates the bind, username lookup, and authentication steps.

The first bind is done with a proxy user ("stuart" in the example
ilibrarian.ini file), whose DN and password are listed in the
ilibrarian.ini file. The search for the username and its DN is done
recursively. A second bind with the found user DN serves as
authentication.

Because the proxy user's DN and password can be considered sensitive
information, the example Apache configs in README.txt have been changed
to forbid access to the ilibrarian.ini file.