Add login authorization with LDAP group membership
Prior to this commit, if a user was found under the LDAP search base,
she was unconditionally allowed to login to the I, Librarian
installation. With this commit, it is possible to restrict logins to
those users who are members of an LDAP group, using the "user_cn_group"
setting in ilibrarian.ini. If the setting is left commented out, the old
behahvior is used instead (i.e., all users under the search base are
eligible for logging in).